Admin Panel Added, About Page Added, Created basic data

+++ pages/api/products/[id].js
@@ -5,8 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
   dbConnect();
+  const token = cookies.token;
 
   if (method === "GET") {
     try {
@@ -18,6 +20,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         // return new
@@ -29,6 +34,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findByIdAndDelete(id);
       res.status(200).json(`The Product with id: ${id} has been deleted.`);
